Renamed agent/proto package to agent/agentpb to prevent package name conflicts
