+++ agent/consul/rpc.go
@@ -10,10 +10,10 @@ import (
 	"time"
 
 	"github.com/armon/go-metrics"
+	"github.com/hashicorp/consul/agent/agentpb"
 	"github.com/hashicorp/consul/agent/consul/state"
 	"github.com/hashicorp/consul/agent/metadata"
 	"github.com/hashicorp/consul/agent/pool"
-	"github.com/hashicorp/consul/agent/proto"
 	"github.com/hashicorp/consul/agent/structs"
 	"github.com/hashicorp/consul/lib"
 	memdb "github.com/hashicorp/go-memdb"
@@ -388,7 +388,7 @@ func (s *Server) raftApplyMsgpack(t structs.MessageType, msg interface{}) (inter
 // raftApplyProtobuf will protobuf encode the request and then run it through raft,
 // then return the FSM response along with any errors.
 func (s *Server) raftApplyProtobuf(t structs.MessageType, msg interface{}) (interface{}, error) {
-	return s.raftApplyWithEncoder(t, msg, proto.EncodeInterface)
+	return s.raftApplyWithEncoder(t, msg, agentpb.EncodeInterface)
 }
 
 // raftApplyWithEncoder is used to encode a message, run it through raft,
